# apps/backends/backend/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Adama Tourism Backend API",
        "version": "1.0.0",
        "description": "Comprehensive API documentation for the Adama Tourism Backend Application."
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key"
            }
        },
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "TOURIST",
                            "RESIDENT"
                        ]
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Profile": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "locale": {
                        "type": "string"
                    }
                }
            },
...
            "Interaction": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "itemId": {
                        "type": "string"
                    }
                }
            },
            "RecordInteractionDto": {
                "type": "object",
                "required": [
                    "itemId",
                    "itemType",
                    "type"
                ],
                "properties": {
                    "itemId": {
                        "type": "string"
                    },
                    "itemType": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "VIEW",
                            "CLICK",
                            "SAVE",
                            "BOOK",
                            "REVIEW",
                            "SHARE"
                        ]
                    },
                    "context": {
                        "type": "object"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "summary": "List users (Admin)",
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/interactions": {
            "get": {
                "summary": "List interactions",
                "tags": [
                    "Interactions"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of interactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Interaction"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Record interaction",
                "tags": [
                    "Interactions"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordInteractionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Interaction recorded"
                    }
                }
            }
        },
...
    }
}